Extract BulkInsertTest and AtomicDirectiveTest, renaming
